Junction pseudostates support completed.
All Junction pseudostate functionalities should work.
Finally removed support for using Statemachine type Behavior as entry
and exit behaviors of States and effect Behaviors of Transitions.
Implementation of this on bare-metal systems would require additional
low-level implicit state machine that would not be visible on the
diagrams and the code would be hard to read by someone not familiar with
the implementation details. Besides, the UML specification is not clear
whether this is allowed or not. The only clue based on 14.2.3.9 in UML
2.5.1 is that a state machine cannot process another event until the
current event is completed. Which might mean that using a Statemachine
as entry, effect, or exit Behaviors would need to violate this rule.
Nevertheless, removing such a possibility from the generator greatly
simplifies it, makes the generated code more readable, and makes no harm
for practical projects. So far, I haven't had a need to implement such a
case in practice. It is better to put explicitly on a diagram some
additional State waiting for processing of another statemachine.
Besides the generator code was cleaned up by adding brackets if the [if]
blocks, since according to the specification they are required.