Added expected fifo indexer.
Added a library for calculating fifo queue indexes to the expected code.
These functions will be used by all event queues, so that the generated
code for calculating and checking the correctness of the tail and head
positions in the event buffers wil not have to be duplicated.
The event queues will access the data in the buffers based on indexes
calculated by this library. Thanks to this the library is very universal
and can be used with any buffer regerdless of the type of data kept
inside.
Apart from this the indexes are easy to check and verify, compared to
pointer based implementations.
Together with the library unit tests are provided in the c_test_expected
project. This project is CMake based project and requires cmake4eclipse
5.1.0 plugin to be installed in the Papyrus from Eclipse Marketplace.
The tests are using the Unity framework.
The last two tests provide examples of using the library with a buffer.