Added event_pool module.
The event_pool module is based on the fifo module, and is used to manage
queues with multiple event types. It allows to set maximum number of
events for each event type separately. This prevents from unintended
blocking of the entire event pool by events of only one type. The
event_pool module was added to prevent code duplication when generating
event queues for each state machine. State machine event queues will be
based on this module. To achieve this, the event queue structs in the
current generated code need to be redesigned.
Added Coverage build configuration to the c_test_expected project.